+++ packages/flutter_meedu/lib/src/state/widgets/watch_filter.dart
@@ -1,10 +1,11 @@
-part of 'consumer_widget.dart';
+import 'package:meedu/provider.dart';
+import 'package:meedu/state.dart';
 
 typedef BuildWhen<S> = bool Function(S prev, S current);
 typedef BuildBySelect<T, S> = S Function(T);
 
 /// class to save a Notifier, the listener and the rebuild function
-class _Target<Notifier, S> extends Provider<Notifier> {
+class Target<Notifier, R> extends Provider<Notifier> {
   /// a SimpleNotifier or a StateNotifier
   final Notifier notifier;
 
@@ -14,23 +15,27 @@ class _Target<Notifier, S> extends Provider<Notifier> {
   /// function to rebuild the Consumer
   void Function()? rebuild;
 
-  _Target(this.notifier);
+  late R selectValue;
+
+  Target(this.notifier);
 }
 
 /// extension for SimpleProvider
 extension SimpleProviderExt<Notifier> on SimpleProvider<Notifier> {
   /// use this method to rebuild your [Consumer] when a value has changed
   /// or you can use a boolean condition. Check the documentation for more info.
-  _Target<Notifier, List> select(BuildBySelect<Notifier, Object?> cb) {
+  Target<Notifier, R> select<R>(BuildBySelect<Notifier, R> cb) {
     // get the  Notifier attached to this SimpleProvider
     final notifier = this.read;
     // get an initial value using the callback
-    Object? prevValue = cb(notifier);
-    final target = _Target<Notifier, List>(notifier);
+    R prevValue = cb(notifier);
+    final target = Target<Notifier, R>(notifier);
+    target.selectValue = prevValue;
 
     // listener with  the logic to rebuild the Consumer
     final listener = (dynamic _) {
       final value = cb(notifier);
+      target.selectValue = value;
       // check if the value has changed
       if (prevValue != value || (value is bool && value)) {
         if (target.rebuild != null) {
@@ -46,8 +51,8 @@ extension SimpleProviderExt<Notifier> on SimpleProvider<Notifier> {
 
   /// use this method to rebuild your [Consumer] using ids (a list of strings)
   /// passed when you call [notify(['id1','id2',...])]
-  _Target<Notifier, List> ids(List<String> Function() cb) {
-    final target = _Target<Notifier, List>(this.read);
+  Target<Notifier, List> ids(List<String> Function() cb) {
+    final target = Target<Notifier, List>(this.read);
     final listener = (dynamic _) {
       // get the ids passed in the notify method
       final listeners = _ as List<String>;
@@ -76,17 +81,18 @@ extension SimpleProviderExt<Notifier> on SimpleProvider<Notifier> {
 }
 
 /// extension for StateProvider
-extension StateProviderExt<Notifier extends StateNotifier<S>, S>
-    on StateProvider<Notifier, S> {
+extension StateProviderExt<Notifier extends StateNotifier<S>, S> on StateProvider<Notifier, S> {
   /// use this method to rebuild your [Consumer] using the previous state and the current
   /// state to return a boolean
-  _Target<Notifier, S> when(BuildWhen<S> cb) {
+  Target<Notifier, bool> when(BuildWhen<S> cb) {
     final notifier = this.read;
-    final target = _Target<Notifier, S>(notifier);
+    final target = Target<Notifier, bool>(notifier);
+    target.selectValue = cb(notifier.state, notifier.state);
 
     final Function(dynamic) listener = (dynamic newState) {
       // rebuild the Consumer using the boolean returned by the callback
       final allowRebuild = cb(notifier.oldState, newState);
+      target.selectValue = allowRebuild;
       if (allowRebuild) {
         if (target.rebuild != null) {
           target.rebuild!();
@@ -99,13 +105,16 @@ extension StateProviderExt<Notifier extends StateNotifier<S>, S>
 
   /// use this method to rebuild your [Consumer] when a value in the state has changed
   /// or you can use a boolean condition. Check the documentation for more info.
-  _Target<Notifier, S> select(BuildBySelect<S, Object?> cb) {
+  Target<Notifier, R> select<R>(BuildBySelect<S, R> cb) {
     final notifier = this.read;
     // get an initial value using the callback
-    Object? prevValue = cb(notifier.state);
-    final target = _Target<Notifier, S>(notifier);
+    R prevValue = cb(notifier.state);
+    final target = Target<Notifier, R>(notifier);
+    target.selectValue = prevValue;
+
     final Function(dynamic) listener = (dynamic newState) {
       final value = cb(notifier.state);
+      target.selectValue = value;
       // check if the value has changed
       if (prevValue != value || (value is bool && value)) {
         if (target.rebuild != null) {