+++ packages/flutter_meedu/lib/state.dart
@@ -2,5 +2,6 @@ library state;
 
 export 'src/state/widgets/consumer_widget/consumer_widget.dart';
 export 'src/state/widgets/consumer_widget/consumer.dart';
+export 'src/state/widgets/watch_filter.dart';
 export 'src/state/listeners/provider_listener.dart';
 export 'src/state/listeners/multi_provider_listener.dart';