+++ website/docusaurus.config.js
@@ -60,7 +60,7 @@ module.exports = {
           lastVersion: "current",
           versions: {
             current: {
-              label: "4.2.1",
+              label: "4.4.0",
               path: "4.x.x",
             },
             "3.x.x": {